Separate triangle degenerate function
